projects/api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
fastapi = "^0.93.0"
uvicorn = "^0.20.0"
pillow = "^9.4.0"
pandas = "^1.5.3"
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
interface = {path = "../interface"}
database = {path = "../database"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
